% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus1.R
\name{minus1}
\alias{minus1}
\title{Minus One}
\usage{
minus1(data, x)
}
\arguments{
\item{data}{the data frame}

\item{x}{the variable (in quotes)}
}
\description{
It takes the variable and minuses one. For internal use in this package.
}
