% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{score.docs}
\alias{score.docs}
\title{Score Docs}
\usage{
score.docs(data, germs, harm, thoughts, symmetry, minus1 = TRUE)
}
\arguments{
\item{data}{the data frame}

\item{germs}{the germs subscale}

\item{harm}{the harm subscale}

\item{thoughts}{the thoughts subscale}

\item{symmetry}{the symmetry subscale}

\item{minus1}{do we want to minus 1 from the raw scores; \code{default = TRUE}}
}
\description{
It scores the DOCS (the Dimensional Obsessive-Compulsive Scale).
}
